policy
Privacy Policy for Omni Craft

Effective Date: December 8, 2023

This Privacy Policy describes how Extrovert GMP ("we," "us," or "our") collects, uses, and discloses information about users of the Omni Craft mobile application ("the App").

Information We Do Not Collect:
We do not collect any personally identifiable information or sensitive user data through the App. This includes but is not limited to names, email addresses, phone numbers, and other personal details.

Age Restriction:
The App is not intended for users under the age of 18. If you are under 18, you may not download, use, or provide information on the App.

Data Collection:
We do not engage in the collection of user data, and the App is designed to operate without requiring the provision of personal information.

Third-Party Services:
The App may integrate third-party services or SDKs (Software Development Kits) for specific functionalities. Please refer to the privacy policies of these third-party services for information on how they handle data.

Security:
We take reasonable measures to protect the information collected through the App from unauthorized access, disclosure, alteration, or destruction. However, please be aware that no method of transmission over the internet or electronic storage is completely secure.

Changes to This Privacy Policy:
We reserve the right to update this Privacy Policy to reflect changes in our practices or for other operational, legal, or regulatory reasons. Any changes will be effective immediately upon posting the updated Privacy Policy on this page.

Contact Us:
If you have any questions about this Privacy Policy, please contact Extrovert GMP at [your contact email address].